LayoutTests/webarchive/loading/test-loading-archive.html

<html><!-- webkit-test-runner [ AlwaysAllowLocalWebarchive=true dumpResourceLoadCallbacks=true ] -->
<script>
    if (window.testRunner) {
        testRunner.dumpFrameLoadCallbacks();
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }
    if (window.internals)
        internals.setAlwaysAllowLocalWebarchive(true);
    function frameLoaded() {
        if (window.testRunner)
            setTimeout(() => testRunner.notifyDone(), 0);
    }
</script>
<body>
    <iframe onload="frameLoaded();" src="resources/helloworld.webarchive"></iframe>
    This tests that doing a "normal load" of a webarchive (not using loadArchive) does not cause a cancelled error to be called.
</body>
</html>
