LayoutTests/platform/wk2/webarchive/loading/test-loading-archive-expected.txt

main frame - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
frame "<!--frame1-->" - didCommitLoadForFrame
frame "<!--frame1-->" - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
 This tests that doing a "normal load" of a webarchive (not using loadArchive) does not cause a cancelled error to be called.
test-loading-archive.html - willSendRequest <NSURLRequest URL test-loading-archive.html, main document URL test-loading-archive.html, http method GET> redirectResponse (null)
test-loading-archive.html - didReceiveResponse <NSURLResponse test-loading-archive.html, http status code 0>
test-loading-archive.html - didFinishLoading
resources/helloworld.webarchive - willSendRequest <NSURLRequest URL resources/helloworld.webarchive, main document URL test-loading-archive.html, http method GET> redirectResponse (null)
resources/helloworld.webarchive - didReceiveResponse <NSURLResponse resources/helloworld.webarchive, http status code 0>
resources/helloworld.webarchive - didFinishLoading
